+++ projects/igniteui-angular/src/lib/directives/button/button.directive.spec.ts
@@ -7,6 +7,14 @@ import { By } from '@angular/platform-browser';
 import { IgxButtonDirective } from './button.directive';
 
 import { configureTestSuite } from '../../test-utils/configure-suite';
+import { IgxIconComponent, IgxIconService } from '../../icon';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { DisplayDensity } from '../../core/density';
+
+const FLAT_RAISED_OUTLINED_BUTTON_COMPACT = 'igx-button--compact';
+const FLAT_RAISED_OUTLINED_BUTTON_COSY = 'igx-button--cosy';
+const FAB_BUTTON_COMPACT = 'igx-button--fab-compact';
+const FAB_BUTTON_COSY = 'igx-button--fab-cosy';
 
 describe('IgxButton', () => {
     configureTestSuite();
@@ -15,8 +23,12 @@ describe('IgxButton', () => {
             declarations: [
                 InitButtonComponent,
                 ButtonWithAttribsComponent,
-                IgxButtonDirective
-            ]
+                ButtonsWithDisplayDensityComponent,
+                IgxButtonDirective,
+                IgxIconComponent
+            ],
+            imports: [HttpClientTestingModule],
+            providers: [IgxIconService]
         })
         .compileComponents();
     }));
@@ -53,6 +65,51 @@ describe('IgxButton', () => {
         expect(button.style.color).toEqual('yellow');
         expect(button.style.background).toEqual('green');
     });
+
+    it('Should apply display density to respective buttons correctly', () => {
+        const fixture = TestBed.createComponent(ButtonsWithDisplayDensityComponent);
+        fixture.detectChanges();
+
+        // Get flat button
+        const flatButton = fixture.componentInstance.flatButton;
+        const flatButtonDOM = fixture.debugElement.query(By.css('.flatBtn'));
+        // Get raised button
+        const raisedButton = fixture.componentInstance.raisedButton;
+        const raisedButtonDOM = fixture.debugElement.query(By.css('.raisedBtn'));
+        // Get outlined button
+        const outlinedButton = fixture.componentInstance.outlinedButton;
+        const outlinedButtonDOM = fixture.debugElement.query(By.css('.outlinedBtn'));
+        // Get fab button
+        const fabButton = fixture.componentInstance.fabButton;
+        const fabButtonDOM = fixture.debugElement.query(By.css('.fabBtn'));
+        // Get icon button
+        const iconButton = fixture.componentInstance.iconButton;
+        const iconButtonDOM = fixture.debugElement.query(By.css('.iconBtn'));
+
+        verifyDisplayDensity(flatButton, flatButtonDOM, 'flat', DisplayDensity.comfortable);
+        verifyDisplayDensity(raisedButton, raisedButtonDOM, 'raised', DisplayDensity.comfortable);
+        verifyDisplayDensity(outlinedButton, outlinedButtonDOM, 'outlined', DisplayDensity.comfortable);
+        verifyDisplayDensity(fabButton, fabButtonDOM, 'fab', DisplayDensity.comfortable);
+        verifyDisplayDensity(iconButton, iconButtonDOM, 'icon', DisplayDensity.comfortable);
+
+        fixture.componentInstance.density = DisplayDensity.compact;
+        fixture.detectChanges();
+
+        verifyDisplayDensity(flatButton, flatButtonDOM, 'flat', DisplayDensity.compact);
+        verifyDisplayDensity(raisedButton, raisedButtonDOM, 'raised', DisplayDensity.compact);
+        verifyDisplayDensity(outlinedButton, outlinedButtonDOM, 'outlined', DisplayDensity.compact);
+        verifyDisplayDensity(fabButton, fabButtonDOM, 'fab', DisplayDensity.compact);
+        verifyDisplayDensity(iconButton, iconButtonDOM, 'icon', DisplayDensity.compact);
+
+        fixture.componentInstance.density = DisplayDensity.cosy;
+        fixture.detectChanges();
+
+        verifyDisplayDensity(flatButton, flatButtonDOM, 'flat', DisplayDensity.cosy);
+        verifyDisplayDensity(raisedButton, raisedButtonDOM, 'raised', DisplayDensity.cosy);
+        verifyDisplayDensity(outlinedButton, outlinedButtonDOM, 'outlined', DisplayDensity.cosy);
+        verifyDisplayDensity(fabButton, fabButtonDOM, 'fab', DisplayDensity.cosy);
+        verifyDisplayDensity(iconButton, iconButtonDOM, 'icon', DisplayDensity.cosy);
+    });
 });
 
 @Component({
@@ -75,4 +132,70 @@ class ButtonWithAttribsComponent {
     public disabled = true;
     public foreground = 'white';
     public background = 'black';
+}
+
+@Component({
+    template:
+    `
+    <button #flat class="flatBtn" igxButton="flat" [displayDensity]="density">Flat</button>
+    <button #raised class="raisedBtn" igxButton="raised" [displayDensity]="density">Raised</button>
+    <button #outlined class="outlinedBtn" igxButton="outlined" [displayDensity]="density">Outlined</button>
+    <button #fab class="fabBtn" igxButton="fab" [displayDensity]="density">
+        <igx-icon fontSet="material">favorite</igx-icon>
+    </button>
+    <button #icon class="iconBtn" igxButton="icon" [displayDensity]="density">
+        <igx-icon fontSet="material">search</igx-icon>
+    </button>
+    `
+})
+class ButtonsWithDisplayDensityComponent {
+    public density: DisplayDensity = DisplayDensity.comfortable;
+
+    @ViewChild('flat', { read: IgxButtonDirective }) flatButton: IgxButtonDirective;
+    @ViewChild('raised', { read: IgxButtonDirective }) raisedButton: IgxButtonDirective;
+    @ViewChild('outlined', { read: IgxButtonDirective }) outlinedButton: IgxButtonDirective;
+    @ViewChild('fab', { read: IgxButtonDirective }) fabButton: IgxButtonDirective;
+    @ViewChild('icon', { read: IgxButtonDirective }) iconButton: IgxButtonDirective;
+}
+
+/**
+ * Verifies the display density of the igxButton based on its type.
+*/
+function verifyDisplayDensity(buttonDirective, buttonDebugEl, buttonType, expectedDisplayDensity: DisplayDensity) {
+    let expectedButtonDensityClass = '';
+
+    switch (expectedDisplayDensity) {
+        case DisplayDensity.compact: {
+            if (buttonType === 'flat' || buttonType === 'raised' || buttonType === 'outlined') {
+                expectedButtonDensityClass = FLAT_RAISED_OUTLINED_BUTTON_COMPACT;
+            } else if (buttonType === 'fab') {
+                expectedButtonDensityClass = FAB_BUTTON_COMPACT;
+            }
+        } break;
+        case DisplayDensity.cosy: {
+            if (buttonType === 'flat' || buttonType === 'raised' || buttonType === 'outlined') {
+                expectedButtonDensityClass = FLAT_RAISED_OUTLINED_BUTTON_COSY;
+            } else if (buttonType === 'fab') {
+                expectedButtonDensityClass = FAB_BUTTON_COSY;
+            }
+        } break;
+        default: break;
+    }
+
+    const buttonNativeElement = buttonDebugEl.nativeElement;
+    if (buttonType === 'icon') {
+        // Icon buttons do not have visual changes for displayDensity.
+        expect(buttonNativeElement.classList.length).toBe(2);
+    } else {
+        if (expectedDisplayDensity === DisplayDensity.comfortable) {
+            // For 'comfortable', the buttons should have no additional css class added.
+            expect(buttonNativeElement.classList.length).toBe(2);
+            expect(buttonNativeElement.classList.contains(expectedButtonDensityClass)).toBe(false, 'Contains density class!');
+        } else {
+            // For 'compact' and 'cosy', the buttons should have an additional css class added.
+            expect(buttonNativeElement.classList.length).toBe(3);
+            expect(buttonNativeElement.classList.contains(expectedButtonDensityClass)).toBe(true, 'Missing density class!');
+        }
+    }
+    expect(buttonDirective.displayDensity).toBe(expectedDisplayDensity);
 }